requirements.txt
Flask
Flask-SQLAlchemy
Flask-JWT-Extended
gunicorn
psycopg2-binary
python-dotenv
Werkzeug
